Add BaseConfigurer
